+++ libs/atomic-state/constants.ts
@@ -1,3 +1,11 @@
+import { createContext } from 'react'
+
 export const $$internal = '$$atom.internal'
 export const $$lifeCycleChannel = '$$atom.lifeCycleChannel'
 export const noop = () => {}
+
+export const defaultContext = Symbol(
+  '$$atom.defaultContext'
+)
+export const RootContext =
+  createContext<any>(defaultContext)
